% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_references.R
\name{activate_references}
\alias{activate_references}
\title{Activate one or more references}
\usage{
activate_references(reference_id, dev = TRUE)
}
\arguments{
\item{reference_id}{String (or integer). One or more 7-digit DataStore reference IDs to be activated}

\item{dev}{Logical. Defaults to TRUE. Should the API call be to the development server (TRUE) or the production server (FALSE)}
}
\value{
dataframe (invisibly)
}
\description{
activate_references takes a list of one or more references, checks their lifecycle status and if they are in Draft form, attempts to activate the reference. If upon attempting activation an error occurs, the user will be notified and the function will move on to the next reference in the list. The function invisibly returns a dataframe where each reference has a row with reference id, initial activation status, current activation status, and an indicator of whether an error was encountered. It is a very wise idea to capture the function output into a variable so that any references that were not activated can be inspected and a re-activation can be attempted. If you forget to do this, it's OK to just re-run the function. It does no harm to attempt to re-activate already active references.
}
\details{
Typical errors include: authentication errors (not on VPN), permissions errors (not a reference owner), or business rules violations (the reference lacks one or more fields required for activation).
}
\examples{
 \dontrun{
 activated_refs <- activate_references(c(1234567, 7654321, 1111111),
                                       dev = TRUE)}
}
