% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_properties.R
\name{add_keyword}
\alias{add_keyword}
\title{Adds a keyword to a list of existing keywords for a DataStore reference}
\usage{
add_keyword(reference_id, keyword, dev = TRUE)
}
\arguments{
\item{reference_id}{String. Integer. A DataStore reference ID to have keywords added to.}

\item{keyword}{String. The keyword to be added}

\item{dev}{Logical. Whether the operation occur on the develoment or production server. Defaults to TRUE (development server)}
}
\value{
NULL (invisibly)
}
\description{
add_another_keyword differs from add_keywords in that it does not replace the existing set of keywords but instead just adds one (or more - haven't tested mulitple keywords yet) to the reference.
}
\examples{
\dontrun{
add_keyword(1234567, "Ants")}
}
