% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_properties.R
\name{add_owners}
\alias{add_owners}
\title{Add one or more owners to a DataStore reference}
\usage{
add_owners(reference_id, owner_list, dev = TRUE)
}
\arguments{
\item{reference_id}{String or Integer. The seven-digit DataStore ID for the reference that owners will be added to}

\item{owner_list}{String or Vector. The owner(s) that will be added to the reference}

\item{dev}{Logical. Whether or not the operations will be performed on the development server. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
Accepts a comma separated list of valid nps (or nps partner) email addresses and adds the people specified via email address as an owners/owners of the specified DataStore reference. Use the email address not the upn/username, e.g. first_last@nps.gov not FMLast@npsg.gov.
}
\examples{
\dontrun{
add_owners(reference_id = 1234567,
           owner_list = "john_doe@nps.gov")
           }
}
