% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_properties.R
\name{add_ref_to_projects}
\alias{add_ref_to_projects}
\title{Adds a product reference to one or more project references}
\usage{
add_ref_to_projects(reference_id, project_id, dev = TRUE)
}
\arguments{
\item{reference_id}{String (or Integer). The seven-digit DataStore reference ID of the product to be added to one or more Project References on DataStore. For instance this could be a Data Package, Web Site, etc.}

\item{project_id}{Integer, String, or Vector. One or more DataStore Project Reference IDs to which the Product reference will be added.}

\item{dev}{Logical. Whether or not the actions will occur on the development server. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
Adds a product reference to one or more project references
}
\examples{
\dontrun{
add_ref_to_projects(reference_id = 1234567, project_id = 7654321)
add_ref_to_projects(reference_id = 1234567, project_id = c(7654321,
                                                           9876543))}
}
