% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_508_format}
\alias{check_508_format}
\title{Checks a column within a supplied file to ascertain whether 508 status has been supplied.}
\usage{
check_508_format(filename, path = getwd())
}
\arguments{
\item{filename}{String. The file to check.}

\item{path}{String. The path to the file. Defaults to the current working directory.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Checks the column files_508_compliant to make sure that all values are either 'yes' or 'no', which indicates the 508 compliant status of any associated files to be uploaded to DataStore. If any values are missing or are not 'yes' or 'no' (where case is ignored), the function throws an error. If all values have either a 'yes' or 'no' value associated with them, the function passes.
}
\examples{
\dontrun{
check_508_format("test_file.txt")}
}
