% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_CUI_label_valid}
\alias{check_CUI_label_valid}
\title{Check a file for valid CUI labels}
\usage{
check_CUI_label_valid(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
Checks the supplied CUI labels against a hard-coded set of accepted CUI markings
}
\examples{
\dontrun{
check_CUI_label_vald(sheet_name = "AudioRecording")}
}
