% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_CUI_license_match}
\alias{check_CUI_license_match}
\title{Checks that license codes are matched with appropriate CUI labels}
\usage{
check_CUI_license_match(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
Checks that license codes are matched with appropriate CUI labels
}
\examples{
\dontrun{
check_CUI_license_match(sheet_name = "AudioRecording")}
}
