% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_dates_past}
\alias{check_dates_past}
\title{Checks that dates in a file occurred in the past.}
\usage{
check_dates_past(filename, path = getwd)
}
\arguments{
\item{filename}{String. Name of the file to check}

\item{path}{String. Path to the file. Defaults to the current working directory}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Throws an error if dates in either the content_begin_date or content_end_date occur after the current system date. If any dates occur in the future the function results in an error. The test passes if all dates occur in the past (or on the present day).
}
\examples{
\dontrun{
check_dates_past("test_file.txt")
}
}
