% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_end_after_start}
\alias{check_end_after_start}
\title{Checks that content_end_dates occur after content_begin_dates in a file.}
\usage{
check_end_after_start(filename, path = getwd)
}
\arguments{
\item{filename}{String. Name of the file to check}

\item{path}{String. Path to the file. Defaults to the current working directory}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Throws an error if any dates in the content_end_date column occur before the date in the same row for the content_begin_date column. Passes if all end dates occur on or after the being date.
}
\examples{
\dontrun{
check_end_after_start("test_file.txt")
}
}
