% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_file_number}
\alias{check_file_number}
\title{Checks whether the total number of file to upload exceeds a threshold}
\usage{
check_file_number(
  filename = "DSbulkUploadR_input.xlsx",
  path = getwd(),
  sheet_name,
  file_number_error = 500
)
}
\arguments{
\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{sheet_name}{String. Name of the excel sheet to check.}

\item{file_number_error}{Integer. Maximum allowable number of files. Defaults to 500.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .xlsx file for data validation. Checks the number of files located in each directory supplied via the column file_path. If the total number of files exceeds the file_number_error value, the function produces an error. If the total number of files exceeds 50\% of the file_number_error value the function produces a warning. Otherwise the function passes.
}
\examples{
\dontrun{
check_file_number("test_file.xlsx")
}
}
