% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_file_size}
\alias{check_file_size}
\title{Checks whether the total file size to upload exceeds some threshold}
\usage{
check_file_size(
  filename = "DSbulkUploadR_input.xlsx",
  path = getwd(),
  sheet_name,
  file_size_error = 100
)
}
\arguments{
\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{sheet_name}{String. Name of the excel sheet to check.}

\item{file_size_error}{Integer. The maximum allowable total file size in GB. Defaults to 100 GB.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .xlsx file for data validation. Checks the file size of files located in each directory supplied via the column file_path. If the total file size exceeds the file_size_error value, the function produces an error. If the total number of files exceeds 50\% of the file_number_error value the function produces a warning. Otherwise the function passes.
}
\examples{
\dontrun{
check_file_size("test_file.xlsx")
}
}
