% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_files_exist}
\alias{check_files_exist}
\title{Checks that each file_path in a file contains files}
\usage{
check_files_exist(filename, path = getwd())
}
\arguments{
\item{filename}{String. The input file to check.}

\item{path}{String. Path to the input file. Defaults to the current working directory}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. For each item in the column file_path, the function checks whether the path given contains file. If the path given does not contain files (or is not a valid path), the function will throw an error and list the bad paths. If all paths contain valid files, the function passes.
}
\examples{
\dontrun{
check_files_exist("test_file.txt")
}
}
