% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_license_type}
\alias{check_license_type}
\title{Checks an file for valid license types}
\usage{
check_license_type(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Checks all values supplied in the license_code column for valid DataStore license codes. Valid codes are as follows: 1) "Creative Commons Zero v1.0 Universal (CC0)". This is the preferred license for all public content. 2) "Creative Commons Attribution 4.0 International", 3) "Creative commons Attribution Non Commercial 4.0 International" (This license may be useful when working with partners external to NPS), 4) "Public Domain", 5) "Unlicensed (not for public dissemination)" (should only be used for restricted content that contains confidential unclassified information - CUI). If all license codes supplied (as integers ranging from 1-5) are valid, the function passes. If any license code is missing or invalid, the function throws an error.
}
\examples{
\dontrun{
check_license_type("test_file.txt")}
}
