% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_owner_email}
\alias{check_owner_email}
\title{Checks owner emails for valid NPS emails}
\usage{
check_owner_email(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
Reads in the specified sheet from an .xlsx file and checks that all email addresses in the owner_email_list column are valid using the AD API.
}
\examples{
\dontrun{
check_owner_email()}
}
