% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_prod_units}
\alias{check_prod_units}
\title{Checks for valid producing unit codes in a file}
\usage{
check_prod_units(filename, path = getwd())
}
\arguments{
\item{filename}{String. Name of file to check}

\item{path}{String. Path to the file. Defaults to the current working directory}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Uses the NPS Unit Service API to check that the producing_units column contains valid NPS unit codes. If invalid producing units are encountered, the function throws an error and prints a list of the invalid units to the console. If all units are valid, the test passes.
}
\examples{
\dontrun{
check_units("test_file.txt")
}
}
