% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_projects_valid}
\alias{check_projects_valid}
\title{Checks for valid projects}
\usage{
check_projects_valid(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name,
  dev = TRUE
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}

\item{dev}{Logical. Whether or not the API calls should be made to the development (TRUE) or production (FALSE) server. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
The function Initiates an API call to datastore to make sure that a project is valid. Potential reasons for check to fail with an error include: The reference ID number does not go to a valid DataStore reference, the valid reference is not a Project type reference, the user not being on the VPN/on an NPS network, the project may not be public, the project may not be active, or the user may not have permissions to access or edit the project.This test passes unconditionally if the reference type is a Project.
}
\examples{
\dontrun{
check_projects_valid(sheet_name = "Script")
}
}
