% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_ref_type}
\alias{check_ref_type}
\title{Checks the reference types in a file are valid Data Store reference types}
\usage{
check_ref_type(filename, path = getwd())
}
\arguments{
\item{filename}{String. Input file to check.}

\item{path}{String. Path to the file. Defaults to the current working directory.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Checks the values supplied in the column reference_type using the DataStore API to determine whether they are valid reference types, as stored in the DataStore backend. Note that the Reference Type listed on the DataStore web page (e.g. Audio Recording) is NOT the correct, valid reference type. In this case the valid reference type would be AudioRecording. If any reference type is invalid, the function throws an error and prints a list of the invalid reference types. If all reference types are valid, the function passes.
}
\examples{
\dontrun{
check_ref_type("test_file.txt")}
}
