% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_refs_identical}
\alias{check_refs_identical}
\title{Check that all reference types are identical}
\usage{
check_refs_identical(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
Runs a check against all values supplied in the reference_type column of the input file. If all values in this column are identical, the test passes. If any are not identical, the test fails.
}
\examples{
\dontrun{check_refs_identical(sheet_name = "Project")}
}
