% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_start_date}
\alias{check_start_date}
\title{Checks the begin_content_date column of an input file for ISO 8601 formatting}
\usage{
check_start_date(filename, path = getwd())
}
\arguments{
\item{filename}{String. name of the file to check}

\item{path}{String. Path to the file. Defaults to the current working directory}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Throws an error if any dates in the content_begin_date column are not in ISO 8601 (yyyy-mm-dd). Passes if all dates in the content_begin_date column are in ISO 8601 format.
}
\examples{
\dontrun{
check_start_date <- ("test_file.txt")}
}
