% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_unique_title}
\alias{check_unique_title}
\title{Checks for duplicated reference titles}
\usage{
check_unique_title(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .xlsx file for data validation. Checks the column title to make sure all values are unique. If any values are duplicated, the function throws an error and returns a list of the duplicated values. If all the values are unique, the function passes.
}
\examples{
\dontrun{
check_unique_title("test_file.xlsx")}
}
