% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_users_AD}
\alias{check_users_AD}
\title{Checks whether a upn supplied is valid.}
\usage{
check_users_AD(filename, path = getwd())
}
\arguments{
\item{filename}{String. File to check.}

\item{path}{String. Path to file. Defaults to the current working directory.}
}
\value{
NULL (invisibly)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Reads in a .xlsx file for data validation. Checks all user names supplied in the column author_upn against Active Directory using a powershell script to verify that the accounts are valid NPS accounts. The function throws an error and lists any invalid accounts if they are encountered. Additionally, if all accounts are valid the function throws and error if any accounts do not have ORCiDs associated with them (and lists those accounts). If all accounts are valid and have ORCiDs, the function passes.
}
\examples{
\dontrun{
check_users_AD("test_file.xlsx")}
}
