% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_uploads.R
\name{create_draft_reference}
\alias{create_draft_reference}
\title{Creates a blank draft reference on DataStore}
\usage{
create_draft_reference(draft_title = "Temp Title", ref_type, dev = TRUE)
}
\arguments{
\item{draft_title}{String. The title for the reference.}

\item{ref_type}{String. The reference type to create.}

\item{dev}{Logical. Should the reference be created on the development server or the production server? Defaults to TRUE}
}
\value{
String. The DataStore reference number.
}
\description{
Creates a blank draft reference on DataStore
}
\examples{
\dontrun{
create_draft_reference(draft_title = "DRAFT TITLE",
                        ref_type = "AudioRecording")}
}
