% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_uploads.R
\name{generate_references}
\alias{generate_references}
\title{Launch bulk reference creation and file upload tool}
\usage{
generate_references(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet,
  max_file_upload = 500,
  max_data_upload = 10,
  data_upload = TRUE,
  max_tries = 3,
  dev = TRUE
)
}
\arguments{
\item{path}{String. Path to the file.}

\item{filename}{String. The name of the file with information on what will be uploaded. Defaults to "DSbulkUploadR_input.xlsx". Must be an xlsx.}

\item{sheet}{String. Name of the sheet within the .xlsx to read data from.}

\item{max_file_upload}{Integer. The maximum allowable number of files to upload. Defaults to 500.}

\item{max_data_upload}{Integer. The maximum allowable amount of data to upload (in GB). Defaults to 100.}

\item{data_upload}{Logical. Defaults to TRUE. To create a bunch of draft reference but not upload any files to them, set the parameter \code{data_upload} to \code{FALSE}.}

\item{max_tries}{Integer. Defaults to 3. The number of attempts that should be made to create and populate any one reference.}

\item{dev}{Logical. Whether the reference creation/file uploads will occur on the development server (TRUE) or the production server (FALSE). Defaults to TRUE.}
}
\value{
Dataframe
}
\description{
The function will first run \code{run_input_validation} on the supplied .txt of information and enforce data validation. All errors must be resolved before the function will proceed.
}
\details{
The function will inform the user of the total number of references to be created, the number of files to be uploaded, and the total volume of data to be uploaded (in GB) and ask the user if they are sure they want to proceed.

The function then creates a draft reference on DataStore for each line in the input .txt and uses the information provided in the .txt to populate the Reference. Finally, all files in the given path for a reference in the .txt will be uploaded to the appropriate reference. All generated references will have the by-or-for-NPS flag set to TRUE. Currently the bulk reference generation tool only supports public references.

The original dataframe generated from the .txt is returned to the user with a single column added: the DataStore reference ID for each newly created reference.
}
\examples{
\dontrun{
generate_references(sheet = "AudioRecording")}
}
