% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_template.R
\name{make_input_template}
\alias{make_input_template}
\title{Write a template input file}
\usage{
make_input_template(reference_type = "AudioRecording", write_path = getwd())
}
\arguments{
\item{reference_type}{String. Case-sensitive DataStore reference type you would like to create an input template for. Defaults to "AudioRecording". Note that the reference_type is a single word and follows the DataStore backend naming conventions, not necessarily the names displayed on the DataStore website (compare AudioRecording with Audio Recording).}

\item{write_path}{String. The path where the file will be written. Defaults to the current working directory.}
}
\value{
dataframe (invisibly)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Writes a template input file (e.g. "AutioRecording_input.xlsx") to the specified path. The template input file can then be edited as needed, saved, and passed to the \code{generate_references} function to create the specified references.

The currently supported reference_types are: AudioRecording.
}
\examples{
\dontrun{
make_input_template()
}
}
