% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_properties.R
\name{remove_content_unit}
\alias{remove_content_unit}
\title{Remove a specific unit code from one or more references}
\usage{
remove_content_unit(reference_id, unit_to_remove, dev = TRUE)
}
\arguments{
\item{reference_id}{String. Integer. Vector. One or more of the seven-digit DataStore reference IDs for the reference(s) that content unit link will be removed.}

\item{unit_to_remove}{String. A single NPS park unit code.}

\item{dev}{Logical. Whether or not the operation will be performed on the development server. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
Remove a specific unit code from one or more references
}
\examples{
\dontrun{
remove_content_unit(2315542, "ROMO")
remove_content_unit(c(2315542, 2315551), "ROMO")}
}
