% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_properties.R
\name{set_content_units}
\alias{set_content_units}
\title{Set content unit links and bounding boxes}
\usage{
set_content_units(reference_id, content_units, dev = TRUE)
}
\arguments{
\item{reference_id}{String. Integer. The seven-digit DataStore ID for the reference that content unit links will be added to.}

\item{content_units}{String. Vector. One or more NPS park unit codes.}

\item{dev}{Logical. Whether or not the operation will be performed on the development server. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
This function takes one or more content unit codes and adds them, along with their bounding boxes, as content unit links to the specified reference on DataStore.
}
\examples{
\dontrun{
set_content_units(reference_id = 1234567,
                  content_units = "ROMO")
set_content_units(reference_id = 1234567,
                  content_units = c("ROMO", "YELL",
                  dev = FALSE))}
}
