% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_uploads.R
\name{upload_files}
\alias{upload_files}
\title{Uploads files of indeterminant size to a DataStore reference}
\usage{
upload_files(
  filename,
  path,
  reference_id,
  is_508 = FALSE,
  chunk_size_mb = 1,
  retry = 1,
  dev = TRUE
)
}
\arguments{
\item{filename}{String. Name of the file to be uploaded.}

\item{path}{String. Location of the file to be uploaded.}

\item{reference_id}{String. The reference number for the DataStore reference the file will be uploaded to.}

\item{is_508}{Logical. Is the file in question 508 compliant? TRUE = Yes, FALSE = No. Defaults to FALSE.}

\item{chunk_size_mb}{Integer. Size of file chunks to be uploaded, in MB}

\item{retry}{Integer. Number of times to retry uploading file chunks if a given chunk fails.}

\item{dev}{Logical. Defaults to TRUE. FALSE means files will be uploaded to the production server. TRUE means files will be uploaded to the development server. Use Dev = TRUE when testing the function.}
}
\value{
List. Of information about the uploaded file.
}
\description{
This file uploader is largely borrowed from the R package NPSdatastore. It will take a file of any size and upload it to a given reference. This involves chunking the file into smaller pieces in order to upload larger files.
}
\examples{
\dontrun{upload_files(filename = "example_file.wav", path = getwd(), reference_id = 1234567, is508 = FALSE, dev = TRUE)}
}
