% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_core_bibliography.R
\name{write_core_bibliography}
\alias{write_core_bibliography}
\title{Write the core bibliography information to DataStore reference landing page}
\usage{
write_core_bibliography(
  reference_id,
  filename,
  row_num,
  path = getwd(),
  dev = TRUE
)
}
\arguments{
\item{reference_id}{Integer. The 7-digit DataStore reference ID where data will be written}

\item{filename}{String. The name of the input.txt file containing data to be added to the reference page}

\item{row_num}{Integer. The row in the file of the input.txt file to be used}

\item{path}{String. Path to input.txt file in \code{filename}. Defaults to \code{getwd()}.}

\item{dev}{Logical. Whether data should be written to the development server or not. Defaults to TRUE.}
}
\value{
NULL (invisibly)
}
\description{
This function takes in a file with the required pre-validated input data (see \code{run_input_validation}) and uses it to populate the core bibliography tab for DataStore references. It currently supports AudioRecording and GenericDocument reference types.
}
\examples{
\dontrun{
write_core_bibliography(reference_id = 1234567,
                        filename = "input.txt",
                        row_num = 1,
                        path = getwd(),
                        dev = TRUE)}
}
