% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_template.R
\name{write_input_template}
\alias{write_input_template}
\title{Write a template input file}
\usage{
write_input_template(write_path = getwd())
}
\arguments{
\item{write_path}{String. The path where the input.xlsx file should be written. Defaults to the current working directory}
}
\value{
NULL (invisibly)
}
\description{
Writes a template .xlsx file ("DSbulkUploadR_input.xlsx") to the specified location.
}
\examples{
\dontrun{
write_input_template()
}
}
