% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_author_email}
\alias{check_author_email}
\title{Verifies author emails supplied in file.}
\usage{
check_author_email(filename, path = getwd())
}
\arguments{
\item{filename}{String. File to check.}

\item{path}{String. Path to the file. Defaults to the current working directory.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Checks that all author emails supplied in the column author_emails are valid NPS emails. Each instance of author emails contain a single email or a comma separated list of emails (e.g. just joe.smith@nps.gov or joe.smith@nps.gov, jane.doe@partner.nps.gov). If any email is not a valid NPS email, the function will throw an error and list the invalid emails (check for typos!). If all emails supplied are valid NPS emails, the function will pass. This function uses an API rather than Active Directory to access lists of valid NPS emails.
}
\examples{
\dontrun{
check_author_email("test_file.txt")}
}
