% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_file_number}
\alias{check_file_number}
\title{Checks whether the total number of file to upload exceeds a threshold}
\usage{
check_file_number(filename, path = getwd(), file_number_error = 500)
}
\arguments{
\item{filename}{String. Name of the file to check.}

\item{path}{String. Path to the file to check. Defaults to the current working directory.}

\item{file_number_error.}{Integer. Maximum allowable number of files. Defaults to 500.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Checks the number of files located in each directory supplied via the column file_path. If the total number of files exceeds the file_number_error value, the function produces an error. If the total number of files exceeds 50\% of the file_number_error value the function produces a warning. Otherwise the function passes.
}
\examples{
\dontrun{
check_file_number("test_file.txt")
}
}
