% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_projects_numeric}
\alias{check_projects_numeric}
\title{Checks that Project IDs are numeric}
\usage{
check_projects_numeric(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name
)
}
\arguments{
\item{path}{String. Path to the file. Defaults to the current working directory.}

\item{filename}{String. Input file to check. Defaults to "DSbulkdUploadR_input.xlsx"}

\item{sheet_name}{String. Name of the excel sheet to check.}
}
\value{
NULL (invisibly)
}
\description{
The function checks that all project IDs are numeric (or NA). If projects are all numeric (or NA), the test passes. Otherwise it fails with a warning.
}
\examples{
\dontrun{
check_projects_numeric(sheet_name = "Script")
}
}
