% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_ref_type_supported}
\alias{check_ref_type_supported}
\title{Check that Reference Types are Supported}
\usage{
check_ref_type_supported(filename, path = getwd())
}
\arguments{
\item{filename}{String. Input file to check.}

\item{path}{}
}
\value{
NULL (invisibly)
}
\description{
Runs a check against a hardcoded list of currently supporte DataStore reference types. If an unsupported reference type is supplied, the function will fail with an error and the error message will return the specific reference type(s) that are unsupported. If the reference types are all supported, the function will pass.  See the \href{https://nationalparkservice.github.io/DSbulkUploadR/articles/02_Generate-the-Input-File.html#create-an-input-template}{package documentation} for a list of currently supported references and information on using them.
}
\examples{
\dontrun{
check_ref_type_supported(filename = "test_file.txt")}
}
