% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation_functions.R
\name{check_unique_title}
\alias{check_unique_title}
\title{Checks for duplicated reference titles}
\usage{
check_unique_title(filename, path = getwd())
}
\arguments{
\item{filename}{String. The file to check.}

\item{path}{String. The path to the file. Defaults to the current working directory.}
}
\value{
NULL (invisibly)
}
\description{
Reads in a .txt file for data validation. Checks the column title to make sure all values are unique. If any values are duplicated, the function throws an error and returns a list of the duplicated values. If all the values are unique, the function passes.
}
\examples{
\dontrun{
check_unique_title("test_file.txt")}
}
