% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSbulkUploadR-package.R
\docType{package}
\name{DSbulkUploadR-package}
\alias{DSbulkUploadR}
\alias{DSbulkUploadR-package}
\title{DSbulkUploadR: Creates, populates, and uploads files to multiple DataStore references}
\description{
Efficiently creates, populates, and uploads files to multiple DataStore references by generating an .xlsx file with predefined fields to be populated by the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DOI-nps/DSbulkUploadR}
  \item \url{https://DOI-nps.github.io/DSbulkUploadR}
  \item \url{https://DOI-nps.github.io/DSbulkUploadR/}
  \item Report bugs at \url{https://github.com/DOI-nps/DSbulkUploadR/issues}
}

}
\author{
\strong{Maintainer}: Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})

Authors:
\itemize{
  \item Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})
}

}
\keyword{internal}
