% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ref_properties.R
\name{remove_editors}
\alias{remove_editors}
\title{Remove editors (owners) from a DataStore reference}
\usage{
remove_editors(reference_id, editor_list, dev = TRUE)
}
\arguments{
\item{reference_id}{String. Contains one 7-digit DataStore reference IDs}

\item{editor_list}{String or list. Contains one or more owner/editor email addresses to be removed from the specified reference.}

\item{dev}{Logical. Defaults to TRUE. Determines whether the operation will be executed on the development server (TRUE) or the production server (FALSE).}
}
\value{
NULL (invisibly)
}
\description{
You must be an owner/editor of the reference to remove owners/editors from the reference. You cannot remove all editors from a reference; all references must have at least one owner/editor. For a given reference ID, you can supply one or more owners/editors to be removed from that reference. Owners/editors to be removed should be supplied as a comma-separated list of email addresses (e.g. john_doe@nps.gov), not upns or AD usernames (e.g. not jdoe@nps.gov).
}
\examples{
\dontrun{
   remove_editors(reference_id = 1234567,
                 editor_list = c("john_doe@nps.gov",
                                "jane_doe@partner.nps.gov"))
   }
}
