% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_core_bibliography.R
\name{write_core_bibliography}
\alias{write_core_bibliography}
\title{Write the core bibliography information to DataStore reference landing page}
\usage{
write_core_bibliography(
  path = getwd(),
  filename = "DSbulkUploadR_input.xlsx",
  sheet_name,
  row_num,
  dev = TRUE
)
}
\arguments{
\item{path}{String. Path to xlsx file in \code{filename}. Defaults to \code{getwd()}.}

\item{filename}{String. The name of the input.xlsx file containing data to be added to the reference page. Defaults to "DSbulkUploadR_input.xlsx".}

\item{row_num}{Integer. The row in the file of the xlsx sheet to be used}

\item{dev}{Logical. Whether data should be written to the development server or not. Defaults to TRUE.}

\item{reference_id}{Integer. The 7-digit DataStore reference ID where data will be written}

\item{sheet}{String. The name of the sheet within the xlsx to get data from.}
}
\value{
NULL (invisibly)
}
\description{
This function takes in a file with the required pre-validated input data (see \code{run_input_validation}) and uses it to populate the core bibliography tab for DataStore references. It currently supports AudioRecording and GenericDocument reference types.
}
\examples{
\dontrun{
write_core_bibliography(reference_id = 1234567,
                        filename = "DSbulkUploadR_input.xlsx",
                        row_num = 1,
                        sheet = "AudioRecording",
                        path = getwd(),
                        dev = TRUE)}
}
